+++ lib/dependencies.sh
@@ -9,8 +9,6 @@ run_if_present() {
 
 install_node_modules() {
   local build_dir=${1:-}
-  local has_prebuild_script=${2:-}
-  local has_postbuild_script=${3:-}
 
   if [ -e $build_dir/package.json ]; then
     cd $build_dir
@@ -23,10 +21,6 @@ install_node_modules() {
       echo "Installing node modules (package.json)"
     fi
     npm install --unsafe-perm --userconfig $build_dir/.npmrc 2>&1
-    if [ -n "$has_postbuild_script" ]; then
-      echo "Running postbuild script"
-      npm run heroku-postbuild
-    fi
   else
     echo "Skipping (no package.json)"
   fi
@@ -34,15 +28,9 @@ install_node_modules() {
 
 rebuild_node_modules() {
   local build_dir=${1:-}
-  local has_prebuild_script=${2:-}
-  local has_postbuild_script=${3:-}
 
   if [ -e $build_dir/package.json ]; then
     cd $build_dir
-    if [ -n "$has_prebuild_script" ]; then
-      echo "Running prebuild script"
-      npm run heroku-prebuild
-    fi
     echo "Rebuilding any native modules"
     npm rebuild 2>&1
     if [ -e $build_dir/npm-shrinkwrap.json ]; then
@@ -51,10 +39,6 @@ rebuild_node_modules() {
       echo "Installing any new modules (package.json)"
     fi
     npm install --unsafe-perm --userconfig $build_dir/.npmrc 2>&1
-    if [ -n "$has_postbuild_script" ]; then
-      echo "Running postbuild script"
-      npm run heroku-postbuild
-    fi
   else
     echo "Skipping (no package.json)"
   fi